+++ x/perp/v2/keeper/dnr.go
@@ -7,11 +7,10 @@ import (
 	"github.com/NibiruChain/collections"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
+	"github.com/NibiruChain/nibiru/x/perp/v2/types"
+
 	"github.com/NibiruChain/nibiru/x/common/asset"
 )
-
-// DnRGCFrequency is the frequency at which the DnR garbage collector runs.
-const DnRGCFrequency = 1000
 
 // IntValueEncoder instructs collections on how to encode a math.Int as a value.
 // TODO: move to collections.
@@ -69,32 +68,54 @@ func (intKeyEncoder) Decode(b []byte) (int, math.Int) {
 
 func (intKeyEncoder) Stringify(key math.Int) string { return key.String() }
 
+// maybeUpdateDnREpoch checks if the current epoch hook call matches the
+// epoch name that targets discounts and rebates, if it does then we simply
+// invoke the StartNewEpoch function to kickstart a new epoch.
+// This method is invoked by the AfterEpochEnd hook.
+func (k Keeper) maybeUpdateDnREpoch(ctx sdk.Context, epochIdentifier string, number uint64) {
+	// if epoch name is empty, we just assume DnR is not enabled.
+	referenceEpochName := k.DnREpochName.GetOr(ctx, "")
+	if referenceEpochName != epochIdentifier {
+		return
+	}
+	// kickstart new epoch
+	k.Logger(ctx).Info("updating dnr epoch", "epochIdentifier", epochIdentifier, "number", number)
+	err := k.StartNewEpoch(ctx, number)
+	if err != nil {
+		// in case of error we panic in this case, because state may have been updated
+		// in a corrupted way.
+		panic(err)
+	}
+}
+
+// StartNewEpoch is called by the epochs hooks when a new 30day epoch starts.
+func (k Keeper) StartNewEpoch(ctx sdk.Context, identifier uint64) error {
+	// set the current epoch
+	k.DnREpoch.Set(ctx, identifier)
+	// we now check the rebates allocated for the previous epoch,
+	// and move them to the escrow such that they can be claimed lazily
+	// by users.
+	previousEpoch := identifier - 1
+	allocationAddr := k.AccountKeeper.GetModuleAddress(types.DNRAllocationModuleAccount)
+	allocationBalance := k.BankKeeper.GetAllBalances(ctx, allocationAddr)
+	err := k.BankKeeper.SendCoinsFromModuleToModule(ctx, types.DNRAllocationModuleAccount, types.DNREscrowModuleAccount, allocationBalance)
+	if err != nil {
+		return err
+	}
+	k.EpochRebateAllocations.Insert(ctx, previousEpoch, types.DNRAllocation{
+		Epoch:  previousEpoch,
+		Amount: allocationBalance,
+	})
+	return nil
+}
+
 // IncreaseTraderVolume adds the volume to the user's volume for the current epoch.
+// It also increases the global volume for the current epoch.
 func (k Keeper) IncreaseTraderVolume(ctx sdk.Context, currentEpoch uint64, user sdk.AccAddress, volume math.Int) {
 	currentVolume := k.TraderVolumes.GetOr(ctx, collections.Join(user, currentEpoch), math.ZeroInt())
 	newVolume := currentVolume.Add(volume)
 	k.TraderVolumes.Insert(ctx, collections.Join(user, currentEpoch), newVolume)
-	k.gcUserVolume(ctx, user, currentEpoch)
-}
-
-// gcUserVolume deletes the un-needed user epochs.
-func (k Keeper) gcUserVolume(ctx sdk.Context, user sdk.AccAddress, currentEpoch uint64) {
-	// we do not want to do this always.
-	if ctx.BlockHeight()%DnRGCFrequency != 0 {
-		return
-	}
-
-	rng := collections.PairRange[sdk.AccAddress, uint64]{}.
-		Prefix(user).                   // only iterate over the user's epochs.
-		EndExclusive(currentEpoch - 1). // we want to preserve current and last epoch, as it's needed to compute DnR rewards.
-		Descending()
-
-	for _, key := range k.TraderVolumes.Iterate(ctx, rng).Keys() {
-		err := k.TraderVolumes.Delete(ctx, key)
-		if err != nil {
-			panic(err)
-		}
-	}
+	k.GlobalVolumes.Insert(ctx, currentEpoch, k.GlobalVolumes.GetOr(ctx, currentEpoch, math.ZeroInt()).Add(volume))
 }
 
 // GetTraderVolumeLastEpoch returns the user's volume for the last epoch.
@@ -139,10 +160,10 @@ func (k Keeper) GetTraderDiscount(ctx sdk.Context, trader sdk.AccAddress, volume
 	return math.LegacyZeroDec(), false
 }
 
-// applyDiscountAndRebate applies the discount and rebate to the given exchange fee ratio.
+// calculateDiscount applies the discount to the given exchange fee ratio.
 // It updates the current epoch trader volume.
 // It returns the new exchange fee ratio.
-func (k Keeper) applyDiscountAndRebate(
+func (k Keeper) calculateDiscount(
 	ctx sdk.Context,
 	_ asset.Pair,
 	trader sdk.AccAddress,
@@ -171,4 +192,69 @@ func (k Keeper) applyDiscountAndRebate(
 	}
 	// return discounted fee ratios
 	return discountedFeeRatio, nil
+}
+
+// WithdrawEpochRebates will withdraw the user's rebates for the given epoch.
+func (k Keeper) WithdrawEpochRebates(ctx sdk.Context, epoch uint64, addr sdk.AccAddress) (withdrawn sdk.Coins, err error) {
+	// get the allocation for the epoch, this also ensures that if the user is trying to withdraw
+	// from current epoch the function immediately fails.
+	allocationCoins, err := k.EpochRebateAllocations.Get(ctx, epoch)
+	if err != nil {
+		return nil, err
+	}
+
+	// compute user weight
+	weight, err := k.computeUserWeight(ctx, addr, epoch)
+	if err != nil {
+		return nil, err
+	}
+	if weight.IsZero() {
+		return sdk.NewCoins(), nil
+	}
+
+	// calculate coins to distribute based on user weight
+	distrCoins := sdk.NewCoins()
+	for _, coin := range allocationCoins.Amount {
+		amt := coin.Amount.ToLegacyDec().Mul(weight).TruncateInt()
+		distrCoins = distrCoins.Add(sdk.NewCoin(coin.Denom, amt))
+	}
+
+	// send money to user from escrow only in case there's anything to distribute.
+	// this should never happen, since we're checking if the user has any weight.
+	if !distrCoins.IsZero() {
+		err = k.BankKeeper.SendCoinsFromModuleToAccount(ctx, types.DNREscrowModuleAccount, addr, distrCoins)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	// garbage collect user volume. This ensures state is not bloated,
+	// and that the user cannot claim from the same allocation twice.
+	return distrCoins, k.TraderVolumes.Delete(ctx, collections.Join(addr, epoch))
+}
+
+// computeUserWeight computes the user's weight for the given epoch.
+func (k Keeper) computeUserWeight(ctx sdk.Context, addr sdk.AccAddress, epoch uint64) (math.LegacyDec, error) {
+	// get user volume for the epoch
+	userVolume := k.TraderVolumes.GetOr(ctx, collections.Join(addr, epoch), math.ZeroInt())
+	if userVolume.IsZero() {
+		return math.LegacyZeroDec(), nil
+	}
+
+	// calculate the user's share
+	globalVolume, err := k.GlobalVolumes.Get(ctx, epoch)
+	if err != nil {
+		return math.LegacyDec{}, err
+	}
+	weight := userVolume.ToLegacyDec().Quo(globalVolume.ToLegacyDec())
+	return weight, nil
+}
+
+// AllocateEpochRebates will allocate the given amount of coins to the current epoch.
+func (k Keeper) AllocateEpochRebates(ctx sdk.Context, sender sdk.AccAddress, amount sdk.Coins) (total sdk.Coins, err error) {
+	err = k.BankKeeper.SendCoinsFromAccountToModule(ctx, sender, types.DNRAllocationModuleAccount, amount)
+	if err != nil {
+		return sdk.Coins{}, err
+	}
+	return k.BankKeeper.GetAllBalances(ctx, k.AccountKeeper.GetModuleAddress(types.DNRAllocationModuleAccount)), nil
 }