+++ x/perp/v2/module/genesis.go
@@ -22,6 +22,8 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 
 	k.DnREpoch.Set(ctx, genState.DnrEpoch)
 
+	k.DnREpochName.Set(ctx, genState.DnrEpochName)
+
 	for _, m := range genState.Markets {
 		k.SaveMarket(ctx, m)
 	}
@@ -77,6 +79,25 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 			panic(err)
 		}
 	}
+
+	for _, globalVolume := range genState.GlobalVolumes {
+		k.GlobalVolumes.Insert(
+			ctx,
+			globalVolume.Epoch,
+			globalVolume.Volume,
+		)
+	}
+
+	for _, rebateAlloc := range genState.RebatesAllocations {
+		k.EpochRebateAllocations.Insert(
+			ctx,
+			rebateAlloc.Epoch,
+			types.DNRAllocation{
+				Epoch:  rebateAlloc.Epoch,
+				Amount: rebateAlloc.Amount,
+			},
+		)
+	}
 }
 
 // ExportGenesis returns the capability module's exported genesis.
@@ -104,6 +125,7 @@ func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 	}
 	genesis.ReserveSnapshots = k.ReserveSnapshots.Iterate(ctx, collections.PairRange[asset.Pair, time.Time]{}).Values()
 	genesis.DnrEpoch = k.DnREpoch.GetOr(ctx, 0)
+	genesis.DnrEpochName = k.DnREpochName.GetOr(ctx, "")
 
 	// export volumes
 	volumes := k.TraderVolumes.Iterate(ctx, collections.PairRange[sdk.AccAddress, uint64]{})
@@ -147,5 +169,19 @@ func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 		panic(err)
 	}
 	genesis.CollateralDenom = collateral
+
+	// export global volumes
+	globalVolumes := k.GlobalVolumes.Iterate(ctx, collections.Range[uint64]{})
+	defer globalVolumes.Close()
+	for ; globalVolumes.Valid(); globalVolumes.Next() {
+		genesis.GlobalVolumes = append(genesis.GlobalVolumes, types.GenesisState_GlobalVolume{
+			Epoch:  globalVolumes.Key(),
+			Volume: globalVolumes.Value(),
+		})
+	}
+
+	// export rebates allocations
+	genesis.RebatesAllocations = k.EpochRebateAllocations.Iterate(ctx, collections.Range[uint64]{}).Values()
+
 	return genesis
 }